// src/lib/types.ts
export interface Beer {
    id: string;
    name: string;
    type: string;
    abv: number;
    origin: string;
    description: string;
    image: string;
}

export interface Wine {
    id: string;
    name: string;
    type: string;
    variety: string;
    vintage: number;
    origin: string;
    description: string;
    image: string;
}